chore(functions): move queue handelr dir
