+++ packages/functions/src/functions/firestore/queues/checkIntegrity.ts
@@ -1,12 +1,12 @@
 import { FirestoreIdData, WatchData } from '@yukukuru/types';
 import * as _ from 'lodash';
-import { updateUserCheckIntegrity } from './firestore/users/integrity';
-import { getRecords } from './firestore/records/getRecords';
-import { updateRecordsStart } from './firestore/records/updateRecordsStart';
-import { getWatches } from './firestore/watches/getWatches';
-import { removeWatches } from './firestore/watches/removeWatches';
-import { getDiffFollowers, getDiffRecords, Diff, DiffWithId } from './diff';
-import { convertRecords } from './convert';
+import { updateUserCheckIntegrity } from '../../../utils/firestore/users/integrity';
+import { getRecords } from '../../../utils/firestore/records/getRecords';
+import { updateRecordsStart } from '../../../utils/firestore/records/updateRecordsStart';
+import { getWatches } from '../../../utils/firestore/watches/getWatches';
+import { removeWatches } from '../../../utils/firestore/watches/removeWatches';
+import { getDiffFollowers, getDiffRecords, Diff, DiffWithId } from '../../../utils/diff';
+import { convertRecords } from '../../../utils/convert';
 
 type Props = {
   uid: string;
